Fix FUN_71015cb6b0 + FUN_7100044dd0 type width — MATCH
FUN_71015cb6b0 (remove): vtable call param was s64, should be int (32-bit)
FUN_7100044dd0: return type was u64, should be u32 (32-bit mov w0)

+++ src/app/fun_med_final_b_002.cpp
@@ -308,7 +308,7 @@ void FUN_7100044c70(long param_1, u64 param_2)
 // 0x7100044dd0  size=64  FUN_7100044dd0
 // Update addr+port if changed; returns 1 on change, 0 if same
 // ---------------------------------------------------------------------------
-u64 FUN_7100044dd0(long param_1, long param_2, s16 param_3)
+u32 FUN_7100044dd0(long param_1, long param_2, s16 param_3)
 {
     if (*(long *)(param_1 + 400) == param_2 && *(s16 *)(param_1 + 0x198) == param_3)
         return 0;
